telegram_bot/requirements.txt
aiogram==3.4.1
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.1
python-dotenv==1.0.0
redis==5.0.1
loguru==0.7.2
